+++ netcdf-examples/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "netcdf-examples"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[features]
+mpi = ["dep:mpi", "netcdf/mpi"]
+
+[dependencies]
+netcdf = { workspace = true }
+mpi = { version = "0.7.0", optional = true }
+mpi-sys = { workspace = true, optional = true }
+ndarray = "0.15.6"
