+++ netcdf/Cargo.toml
@@ -32,7 +32,6 @@ mpi-sys = { workspace = true, optional = true }
 
 [dev-dependencies]
 clap = { version = "4.5.1", features = ["derive"] }
-mpi = "0.7.0"
 tempfile = "3.1.0"
 
 [package.metadata.docs.rs]
